+++ CHANGELOG.md
@@ -5,14 +5,14 @@
 **@publicodes/core**
 
 - Ajoute une méthode `safeGetSituation` à l'Engine.
-- Ajoute une option `filterSituation: boolean` à la méthode `setSituation` pour permettre de filter les règles invalides de la situation dans 2 cas :
+- Ajoute une option `shouldFilterSituation: boolean` à la méthode `setSituation` pour permettre de filter les règles invalides de la situation dans 2 cas :
 
   - La règle n'est pas présente dans la base de règle
   - La réponse donnée via un mécanisme "une possibilité" n'est pas une option de réponse
 
   Désormais, si l'option est activée, l'évaluation de cassera pas, seule un warning sera levé.
 
-- Ajoute une option `useSafeGetSituation` à l'Engine pour utiliser activer l'option `filterSituation` par défaut.
+- Ajoute une option `useSafeGetSituation` à l'Engine pour utiliser activer l'option `shouldFilterSituation` par défaut.
 
 ## 1.2.0